requirements.txt
autopep8==1.4.4
pep8==1.7.1
pkg-resources==0.0.0
pycodestyle==2.5.0
RPi.GPIO==0.7.0
